Fix mkt-banner not auto-dismissing (bug 1172707)
There was a regression when the close animation was added that caused
the mkt-banner to not be dismissed when it wasn't visble since the
transitionend event would not fire. The banner will now always be
dismissed after 500ms and it will be immediately dismissed when it is
first created if it has been dismissed and dismiss="remember" or
dismiss="session".